make Transaction mod revision 64 bit

+++ etcd/v3/Transaction.hpp
@@ -36,7 +36,7 @@ public:
 
 	void init_compare(CompareResult, CompareTarget);
 	void init_compare(std::string const &old_value, CompareResult, CompareTarget);
-	void init_compare(int old_value, CompareResult, CompareTarget);
+	void init_compare(int64_t old_value, CompareResult, CompareTarget);
 
 	void setup_basic_failure_operation(std::string const &key);
 	void setup_set_failure_operation(std::string const &key, std::string const &value, int64_t leaseid);
